+++ core/mybatis-generator-systests-ibatis2-java2/src/test/java/mbg/test/ib2j2/flat/FlatJava2Test.java
@@ -248,7 +248,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(new Integer(5));
 
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -417,7 +417,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -1124,7 +1124,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example = new PkfieldsExample();
@@ -1405,7 +1405,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(new Integer(4));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example = new PkblobsExample();
@@ -1777,7 +1777,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(new Integer(3));
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example = new PkfieldsblobsExample();
@@ -1963,7 +1963,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example = new FieldsblobsExample();
@@ -2971,7 +2971,7 @@ public class FlatJava2Test extends AbstractFlatJava2Test {
 
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
