+++ core/mybatis-generator-systests-ibatis2-java5/src/test/java/mbg/test/ib2j5/hierarchical/HierarchicalJava5Test.java
@@ -226,7 +226,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(5);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -404,7 +404,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(4);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -1129,7 +1129,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1431,7 +1431,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(4);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1831,7 +1831,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(3);
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -2024,7 +2024,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -3056,7 +3056,7 @@ public class HierarchicalJava5Test extends AbstractHierarchicalJava5Test {
 
             AwfulTableExample example = new AwfulTableExample();
             example.createCriteria().andEMailLike("fred@%");
-            int rows = dao.countByExample(example);
+            long rows = dao.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
